feat(plans): move plans page to lib and use in apps
