+++ web/apps/admin/src/pages/plans/PlansPage.tsx
@@ -1,0 +1,14 @@
+import { PlansView } from "@raystack/frontier/admin";
+import { useNavigate, useParams } from "react-router-dom";
+
+export function PlansPage() {
+  const { planId } = useParams();
+  const navigate = useNavigate();
+
+  return (
+    <PlansView
+      selectedPlanId={planId}
+      onCloseDetail={() => navigate("/plans")}
+    />
+  );
+}
