+++ web/lib/admin/views/plans/header.tsx
@@ -1,5 +1,5 @@
 import { DataTable } from "@raystack/apsara";
-import PageHeader from "~/components/page-header";
+import { PageHeader } from "../../components/PageHeader";
 import styles from "./plans.module.css";
 
 export const PlanHeader = ({ header }: any) => {
