+++ web/lib/admin/views/plans/plans.module.css
